requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
cloudpickle==3.1.1
fastapi==0.115.13
h11==0.16.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.5.1
llvmlite==0.44.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
